AI-MoM-Generator/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# AI/ML Processing
openai==1.3.7
pyannote.audio==3.1.1
torch==2.1.1
torchaudio==2.1.1

# Audio/Video Processing
ffmpeg-python==0.2.0

# Email
sendgrid==6.10.0

# Background Jobs
celery==5.3.4
redis==5.0.1

# Utilities
requests==2.31.0
Pillow==10.1.0
